+++ src/run/emulator.rs
@@ -4,7 +4,7 @@
 
 //!Provides utilities for emulation of AVM bytecode.
 
-use super::runtime_env::RuntimeEnvironment;
+use super::RuntimeEnvironment;
 use crate::compile::{CompileError, DebugInfo};
 use crate::link::LinkedProgram;
 use crate::mavm::{AVMOpcode, Buffer, CodePt, Instruction, Value};
